PR feedback based on Evan's comments
